requirements.txt
anthropic
diskcache
pypdf
openreview-py
litellm
ipykernel
git+https://github.com/nityathakkar/textgrad.git@main